+++ orgflow/src/lib/core/mod.rs
@@ -1,5 +1,5 @@
 mod dates;
 pub mod note;
 mod priority;
-mod tags;
+pub mod tags;
 pub mod task;